Delete sync provider POST surface and small protocol dead ends
Remove production code with no live callers and move the deterministic
transport fakes into test infrastructure:

- headers.py: drop require_header_partition and the two header-name
  frozensets; headers_test.py now asserts the header sets inline. The
  gateway only builds headers, it never re-validates a partition.
- streaming.py: drop encode_chat_events and encode_responses_events;
  streaming_test.py and sdk_harness_test.py carry a local _encode
  helper instead. No production caller existed.
- errors.py: narrow unsupported_field to the parameter-only shape; the
  single caller passes no capability kwarg and the public
  unsupported_capability wire code stays reachable through
  public_failure_error.
- transport.py: delete post_json and the sync POST methods on
  JsonHttpTransport and HttpxJsonTransport; the sync path is GET-only
  (provider listing). Relocate ScriptedJsonTransport (now GET-only)
  and RecordedRequest to a new wmo/runtime/models/conftest.py.
- async_transport.py: delete get_json_async, SyncJsonTransportAdapter,
  and as_async_transport; base/openai/azure client transports narrow
  to AsyncJsonHttpTransport and base defaults to
  HttpxAsyncJsonTransport(). registry.ProviderTransport narrows the
  same way; no production caller injected a sync transport. Relocate
  ScriptedAsyncJsonTransport to the shared conftest and migrate every
  client test to it (base_test's timeout-recording subclass is
  replaced by the fake's native timeouts list).
- protocol.py: delete SyncModelClientAdapter, AsyncCompletedModelClient,
  and BoundedSyncModelClientAdapter.complete_async/_release_permit.
  The class, __init__, complete, _acquire, and _permits stay because
  BoundedBedrockClient.stream inherits them; stream releases permits
  through its own callbacks, never through _release_permit.
- providers/__init__.py drops the two deleted protocol exports.
- pyproject.toml wheel exclude generalizes to **/conftest.py so the
  new test-only module never ships.

ruff, ty, and the models/openai_protocol/gateway test suites are green.
