+++ wmo/runtime/models/providers/transport_test.py
@@ -5,11 +5,11 @@ from __future__ import annotations
 import pytest
 
 from wmo.common.core.artifacts import JsonObject
+from wmo.runtime.models.conftest import ScriptedJsonTransport
 from wmo.runtime.models.providers.transport import (
     JsonHttpResponse,
     ProviderTransportError,
     RetryPolicy,
-    ScriptedJsonTransport,
     classify_retry,
     get_json,
     run_with_retry,
